Separate all the classes inheriting from Hamiltonian into a separate module
